requirements.txt
flask==3.0.0
python-dotenv==1.0.0
twilio==8.10.0
openai==1.3.0
sendgrid==6.10.0
python-dateutil==2.8.2